✏CHANGE: Move Function classes to submodule
