+++ equator/lib/functions/__init__.py
@@ -1,0 +1,7 @@
+"""Functions module
+
+Contains definitions for how to execute functions
+"""
+
+from .function import Function
+from .negate_function import NegateFunction
